1、修改枚举类型 2、def.h增加命名空间

+++ TcpCommucation/ConnDisconnectEvt.cpp
@@ -8,7 +8,7 @@ namespace tc
 	ConnDisconnectEvt::ConnDisconnectEvt(TcpService* pSrv, SOCKET sendrecv/* = NULL*/)
 		:TcpEvt(pSrv, sendrecv)
 	{
-		evt = static_cast<int>(ETcpEvt::ConnDisconnect);
+		evt = ETcpEvt::ConnDisconnect;
 	}
 
 	ConnDisconnectEvt::~ConnDisconnectEvt()
